requirements.txt
mss==9.0.1
Pillow==10.2.0
requests==2.31.0

# Optional: S3/MinIO support
boto3==1.34.34

# Optional: SFTP support
paramiko==3.4.0

# Optional: WebDAV support
webdavclient3==3.14.6

